+++ R/corrplot.R
@@ -42,14 +42,15 @@
 #'   are still distributed uniformly in [-1, 1], it only affect the display
 #'   on color-legend.
 #'
+#' @param is.corr Logical, whether the input matrix is a correlation matrix or
+#'   not. We can visualize the non-correlation matrix by setting
+#'   \code{is.corr = FALSE}.
+#'
 #'
 #' @param bg The background color.
 #'
 #' @param title Character, title of the graph.
 #'
-#' @param is.corr Logical, whether the input matrix is a correlation matrix or
-#'   not. We can visualize the non-correlation matrix by setting
-#'   \code{is.corr = FALSE}.
 #'
 #' @param add Logical, if \code{TRUE}, the graph is added to an existing plot,
 #'   otherwise a new plot will be created.
@@ -100,6 +101,7 @@
 #'
 #' @param rect.lwd Numeric, line width for borders for rectangle border(s), only
 #'   valid when \code{addrect} is equal or greater than 1.
+#'
 #'
 #' @param tl.pos Character or logical, position of text labels. If character, it
 #'   must be one of \code{'lt'}, \code{'ld'}, \code{'td'}, \code{'d'} or
@@ -161,6 +163,12 @@
 #' @param shade.lwd Numeric, the line width of shade.
 #'
 #' @param shade.col The color of shade line.
+#'
+#' @param transKeepSign Logical, whether or not to keep matrix values' sign when
+#'   transforming non-corr matrix for plotting.
+#'   Only valid when \code{is.corr = FALSE}. The default value is \code{TRUE}.
+#'
+#'   NOTE: If \code{FALSE},the non-corr matrix will be
 #'
 #' @param p.mat Matrix of p-value, if \code{NULL}, parameter \code{sig.level},
 #'   \code{insig}, \code{pch}, \code{pch.col}, \code{pch.cex} are invalid.
@@ -259,8 +267,8 @@
 #' @export
 corrplot = function(corr,
   method = c('circle', 'square', 'ellipse', 'number', 'shade', 'color', 'pie'),
-  type = c('full', 'lower', 'upper'), col = NULL, col.lim = NULL, bg = 'white',
-  title = '', is.corr = TRUE, add = FALSE,   diag = TRUE, outline = FALSE,
+  type = c('full', 'lower', 'upper'), col = NULL, col.lim = NULL, is.corr = TRUE,
+  bg = 'white',   title = '', add = FALSE, diag = TRUE, outline = FALSE,
   mar = c(0, 0, 0, 0),
 
   addgrid.col = NULL, addCoef.col = NULL, addCoefasPercent = FALSE,
@@ -280,6 +288,8 @@ corrplot = function(corr,
 
   addshade = c('negative', 'positive', 'all'),
   shade.lwd = 1, shade.col = 'white',
+
+  transKeepSign = TRUE,
 
   p.mat = NULL, sig.level = 0.05,
   insig = c('pch', 'p-value', 'blank', 'n', 'label_sig'),
@@ -316,6 +326,10 @@ corrplot = function(corr,
   # select grid color automatically if not specified
   if (is.null(addgrid.col)) {
     addgrid.col = switch(method, color = NA, shade = NA, 'grey')
+  }
+
+  if(!is.corr & !transKeepSign & method %in% c('circle', 'square', 'ellipse', 'shade', 'pie')) {
+    stop("method should not be in c('circle', 'square', 'ellipse', 'shade', 'pie') when transKeepSign = FALSE")
   }
 
   # Issue #142
@@ -380,15 +394,14 @@ corrplot = function(corr,
       warning('col.lim interval too wide, please set a suitable value')
     }
 
-    # all negative or positive, trans to [0, 1]
-    if (c_max <= 0 | c_min>=0) {
+    # all negative or positive or NOT transkeepSign, trans to [0, 1]
+    if (c_max <= 0 | c_min>=0 | !transKeepSign) {
       intercept = - col.lim[1]
       zoom = 1 / (diff(col.lim))
 
-
-      if(col.lim[1] * col.lim[2] < 0) {
-        warning('col.lim interval not suitable to the matrix')
-      }
+      #if(col.lim[1] * col.lim[2] < 0) {
+      #  warning('col.lim interval not suitable to the matrix')
+      #}
 
     }
 
@@ -866,9 +879,20 @@ corrplot = function(corr,
   ## add numbers
   if (!is.null(addCoef.col) && method != 'number') {
     text(Pos[, 1], Pos[, 2],  col = addCoef.col,
-         labels = round((DAT - int) * ifelse(addCoefasPercent, 100, 1) / zoom,
-                        number.digits),
+         labels = format(
+		   round(
+		     (DAT - int) * ifelse(addCoefasPercent, 100, 1) / zoom, 
+			 number.digits
+		   ), 
+		   nsmall = number.digits
+		 ),
          cex = number.cex, font = number.font)
+  }
+
+
+  if (!is.null(p.mat)) {
+    pos.pNew  = getPos.Dat(p.mat)[[1]]
+    pNew      = getPos.Dat(p.mat)[[2]]
   }
 
 
@@ -883,10 +907,6 @@ corrplot = function(corr,
         warning('p.mat and corr may be not paired, their rownames and colnames are not totally same!')
       }
     }
-
-
-    pos.pNew  = getPos.Dat(p.mat)[[1]]
-    pNew      = getPos.Dat(p.mat)[[2]]
 
     if (insig == 'label_sig') {